+++ source/Internal/Utility/ContextInterface.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace OxidEsales\EshopCommunity\Internal\Utility;
+
+/**
+ * @internal
+ */
+interface ContextInterface
+{
+    /**
+     * @return string
+     */
+    public function getLogLevel();
+
+    /**
+     * @return string
+     */
+    public function getLogFilePath();
+}